SerrayClient/src/test/java/com/nhnacademy/serrayclient/controller/CommentControllerTest.java
package com.nhnacademy.serrayclient.controller;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.anyInt;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.atLeastOnce;
import static org.mockito.Mockito.doNothing;
import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static reactor.core.publisher.Mono.when;

import com.nhnacademy.serrayclient.data.vo.CommentForm;
import com.nhnacademy.serrayclient.service.CommentService;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.web.servlet.MockMvc;

@WithMockUser
@AutoConfigureMockMvc(addFilters = false)
@WebMvcTest(CommentController.class)
class CommentControllerTest {

    @Autowired
    MockMvc mockMvc;

    @MockBean
    CommentService service;

    @DisplayName("코멘트 수정")
    @Test
    void commentModify() throws Exception {

        doNothing().when(service).modifyComment(anyInt(), anyString());

        mockMvc.perform(post("/comment/modify")
            .param("taskNo", "1")
            .param("commentNo", "1")
            .param("projectNo", "1")
            .param("content", "Hello"))
            .andExpect(status().is3xxRedirection())
            .andExpect(result -> assertThat(result.getResponse().getRedirectedUrl())
                .isEqualTo("/task/detail?taskNo=1&projectNo=1"));

        verify(service, atLeastOnce()).modifyComment(anyInt(), anyString());
    }
}